requirements.txt
# Core dependencies
openai==1.54.0
python-dotenv==1.0.0
requests==2.31.0

# Web scraping and search
firecrawl-py==0.0.16
perplexity-sdk==0.1.0

# Database
sqlalchemy==2.0.23

# UI
streamlit==1.39.0
plotly==5.18.0
pandas==2.1.4

# Export
fpdf==1.7.2
reportlab==4.0.7

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
